requirements.txt
Flask
openai-whisper
pydub
google-generativeai
python-dotenv